Extracted token parser
