+++ src/parser/filter.rs
@@ -1,4 +1,4 @@
-use nom::{character::complete::multispace0, multi::many0, sequence, IResult};
+use nom::IResult;
 
 use crate::app::config::Config;
